+++ docs/conf.py
@@ -22,8 +22,8 @@ copyright = '2019, RTE France'
 author = 'Benjamin Donnot'
 
 # The full version, including alpha/beta/rc tags
-release = '0.3.7'
-version = '0.3'
+release = '0.4.0'
+version = '0.4'
 
 
 # -- General configuration ---------------------------------------------------