+++ grid2op/__init__.py
@@ -6,7 +6,7 @@ Document will be made later on.
 import os
 import pkg_resources
 
-__version__ = '0.3.7'
+__version__ = '0.4.0'
 
 __all__ = ['Action', "BackendPandaPower", "Agent", "Backend", "ChronicsHandler", "Environment", "Exceptions",
            "Observation", "Parameters", "GameRules", "Reward", "Runner", "main", "Utils", "PlotPlotly"]